Refactor : commentdto 구조변경
