+++ src/test/java/io/powerrangers/backend/service/CommentServiceTests.java
@@ -2,9 +2,9 @@ package io.powerrangers.backend.service;
 
 import io.powerrangers.backend.dto.TaskScope;
 import io.powerrangers.backend.dto.TaskStatus;
-import io.powerrangers.backend.dto.comment.CommentCreateRequestDto;
-import io.powerrangers.backend.dto.comment.CommentResponseDto;
-import io.powerrangers.backend.dto.comment.CommentUpdateRequestDto;
+import io.powerrangers.backend.dto.CommentCreateRequestDto;
+import io.powerrangers.backend.dto.CommentResponseDto;
+import io.powerrangers.backend.dto.CommentUpdateRequestDto;
 import io.powerrangers.backend.entity.Comment;
 import io.powerrangers.backend.entity.Task;
 import io.powerrangers.backend.entity.User;
@@ -12,7 +12,6 @@ import io.powerrangers.backend.dao.CommentRepository;
 import io.powerrangers.backend.dao.TaskRepository;
 import io.powerrangers.backend.dao.UserRepository;
 import io.powerrangers.backend.exception.CustomException;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
